[feat/#175] :recycle: Refactor: 퀴즈방 생성과 방 상태 도메인 생성을 분리
