+++ chatty-be/src/main/java/com/chatty/chatty/quizroom/service/GameService.java
@@ -1,7 +1,8 @@
 package com.chatty.chatty.quizroom.service;
 
+import com.chatty.chatty.quizroom.controller.dto.PlayersStatusDTO;
 import com.chatty.chatty.quizroom.domain.PlayersStatus;
-import com.chatty.chatty.quizroom.repository.RoomUsersStatusRepository;
+import com.chatty.chatty.quizroom.repository.PlayersStatusRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -9,9 +10,13 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class GameService {
 
-    private final RoomUsersStatusRepository roomUsersStatusRepository;
+    private final PlayersStatusRepository playersStatusRepository;
 
-    public PlayersStatus joinRoom(Long roomId, Long userId) {
-        return roomUsersStatusRepository.addUserToRoom(roomId, userId);
+    public PlayersStatusDTO joinRoom(Long roomId, Long userId) {
+        PlayersStatus playersStatus = playersStatusRepository.saveUserToRoom(roomId, userId);
+        return PlayersStatusDTO.builder()
+                .roomId(roomId)
+                .playerStatuses(playersStatus.playerStatusSet())
+                .build();
     }
 }
